refactored components, created custom hook for mutation, and futhur normalized the components
