+++ src/components/Search.jsx
@@ -1,0 +1,16 @@
+import React from "react";
+
+const Search = () => {
+  return (
+    <div className="mb-3">
+      <input
+        type="text"
+        className="form-control"
+        placeholder="Filter Contacts..."
+        id="exampleInputPassword1"
+      />
+    </div>
+  );
+};
+
+export default Search;
